fix: db transaction issue, retry pending
If some issue error occurs it should exist a
`SageX3TransactionInformation` object.
fixes #265
related to #261
Also fix django command that retry pending transactions.
Refactor billing test
